requirements.txt
streamlit
python-dotenv
openai
Flask
requests
